+++ examples/java/CIGHost.java
@@ -7,10 +7,10 @@ public class CIGHost {
 
     public static void main (String[] args) {
 
-        DoomGame game = new DoomGame();
-
         System.out.println("\n\nCIG HOST EXAMPLE\n");
 
+
+        DoomGame game = new DoomGame();
 
         // Use CIG example config or Your own.
         game.loadConfig("../../scenarios/cig.cfg");
@@ -24,13 +24,15 @@ public class CIGHost {
 
         // Host game with options that will be used in the competition.
         game.addGameArgs("-host 8 "                // This machine will function as a host for a multiplayer game with this many players (including this machine). It will wait for other machines to connect using the -join parameter and then start the game when everyone is connected.
-                        +"-deathmatch "            // Deathmatch rules are used for the game.
-                        +"+timelimit 10.0 "        // The game (episode) will end after this many minutes have elapsed.
-                        +"+sv_forcerespawn 1 "     // Players will respawn automatically after they die.
-                        +"+sv_noautoaim 1 "        // Autoaim is disabled for all players.
-                        +"+sv_respawnprotect 1 "   // Players will be invulnerable for two second after spawning.
-                        +"+sv_spawnfarthest 1 "    // Players will be spawned as far as possible from any other players.
-                        +"+viz_nocheat 1");        // Disables depth and labels buffer and the ability to use commands that could interfere with multiplayer game.
+                        + "-deathmatch "           // Deathmatch rules are used for the game.
+                        + "+timelimit 10.0 "       // The game (episode) will end after this many minutes have elapsed.
+                        + "+sv_forcerespawn 1 "    // Players will respawn automatically after they die.
+                        + "+sv_noautoaim 1 "       // Autoaim is disabled for all players.
+                        + "+sv_respawnprotect 1 "  // Players will be invulnerable for two second after spawning.
+                        + "+sv_spawnfarthest 1 "   // Players will be spawned as far as possible from any other players.
+                        + "+sv_nocrouch 1 "        // Disables crouching.
+                        + "+viz_respawn_delay 10 " // Sets delay between respanws (in seconds).
+                        + "+viz_nocheat 1");       // Disables depth and labels buffer and the ability to use commands that could interfere with multiplayer game.
 
         // Name your agent and select color
         // colors: 0 - green, 1 - gray, 2 - brown, 3 - red, 4 - light gray, 5 - light brown, 6 - light red, 7 - light blue