+++ examples/lua/cig_bots.lua
@@ -16,14 +16,14 @@ game:setDoomMap("map01") -- Limited deathmatch.
 
 -- Start multiplayer game only with your AI (with options that will be used in the competition, details in cig_host example).
 game:addGameArgs("-host 1 -deathmatch +timelimit 1 " ..
-                 "+sv_forcerespawn 1 +sv_noautoaim 1 " ..
-                 "+sv_respawnprotect 1 +sv_spawnfarthest 1")
+                 "+sv_forcerespawn 1 +sv_noautoaim 1 +sv_respawnprotect 1 +sv_spawnfarthest 1 +sv_nocrouch 1 " ..
+                 "+viz_respawn_delay 10 +viz_nocheat 1")
 
 -- Name your agent and select color
 -- colors: 0 - green, 1 - gray, 2 - brown, 3 - red, 4 - light gray, 5 - light brown, 6 - light red, 7 - light blue
 game:addGameArgs("+name AI +colorset 0")
 
-game:setMode(vizdoom.Mode.PLAYER)
+game:setMode(vizdoom.Mode.SPECTATOR)
 --game:setWindowVisible(false)
 
 game:init();
@@ -71,7 +71,7 @@ for i = 1, episodes do
             game:respawnPlayer()
         end
 
-        print("Frags:", game:getGameVariable(vizdoom.GameVariable.FRAGCOUNT))
+        --print("Frags:", game:getGameVariable(vizdoom.GameVariable.FRAGCOUNT))
     end
 
     print("Episode finished.")